+++ lab6/mem2/mem.h
@@ -2,8 +2,8 @@
 #ifndef MEM_H
 #define MEM_H
 
-extern int heap;
-extern int maxheap;
+extern long long int heap;
+extern long long int maxheap;
 
 
 void *myalloc(int size);
